pyproject.toml
[project]
name = "ca-slack-block-kit"
version = "0.1.0"
description = "A collection of simple, pre-configured Slack block kit components to use with Slack notification messages."
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.13.0",
    "slack-sdk>=3.38.0",
    "ty>=0.0.1a21",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "ca_slack_block_kit"
module-root = "."

[tool.uv]
required-version = ">=0.8.4"

[tool.ruff]
target-version = 'py313'
line-length = 95
indent-width = 4

[tool.ruff.lint]
select = [
    "RUF", # Ruff specific rules
    "N",   # pep8-naming
    "I",   # isort
    "F",   # pyflakes
    "UP",  # pyupgrade
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "ANN", # flake8-annotations
    "S",   # flake8-bandit
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]

ignore = [
    "E501", # line-too-long
    "T201", # print
    "S101", # use of assert detected
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ty.environment]
root = ["ca_slack_block_kit"]
python = "./.venv"
